GH-8323 move cypress tests
